+++ src/depyty/cli/errors.py
@@ -1,2 +1,0 @@
-class PrintUsage(Exception):
-    """Can be used to instruct the CLI to print its usage."""