+++ src/depyty/introspection/external.py
@@ -3,17 +3,12 @@ from inspect import getsource
 from subprocess import run
 
 from depyty.introspection import script as script_module
-from depyty.introspection.script import Module
 
 
-def get_available_modules_by_name_standalone(
-    python_path: str,
-) -> dict[str, Module]:
+def introspect(python_path: str) -> script_module.IntrospectionResult:
     script = getsource(script_module)
     script += "\n\n"
-    script += "import json"
-    script += "\n\n"
-    script += "print(json.dumps([module.to_json_dict() for module in get_available_modules_by_name().values()]))"
+    script += "print_serialized_introspection_result()"
 
     # This is a _bit_ hacky, since the buffers for stdout could overflow, but
     # let's see how long this works
@@ -21,9 +16,31 @@ def get_available_modules_by_name_standalone(
         [python_path, "-"], input=script.encode(), capture_output=True, check=True
     )
 
-    modules: dict[str, Module] = {}
-    for serialized_module in json.loads(process.stdout):
-        module = Module.from_json_dict(serialized_module)
-        modules[module.name] = module
+    parsed_stdout = json.loads(process.stdout)
 
-    return modules
+    if not isinstance(parsed_stdout, dict):
+        raise ValueError("Expected introspection to yield a dict")
+
+    top_level_stdlib_module_names = parsed_stdout.get("top_level_stdlib_module_names")
+    if not isinstance(top_level_stdlib_module_names, list):
+        raise ValueError(
+            "Expected introspection to yield a dict where 'top_level_stdlib_module_names' is a list"
+        )
+
+    distribution_names = parsed_stdout.get("distribution_names")
+    if not isinstance(distribution_names, list):
+        raise ValueError(
+            "Expected introspection to yield a dict where 'distribution_names' is a list"
+        )
+
+    path = parsed_stdout.get("path")
+    if not isinstance(path, list):
+        raise ValueError(
+            "Expected introspection to yield a dict where 'path' is a list"
+        )
+
+    return script_module.IntrospectionResult(
+        top_level_stdlib_module_names=top_level_stdlib_module_names,
+        distribution_names=distribution_names,
+        path=path,
+    )