+++ src/depyty/introspection/script.py
@@ -8,82 +8,35 @@ NOTE: This file may not depend upon anything that is not available in the stdlib
       packages/modules are available in them.
 """
 
-import os
+import json
 import sys
-import sysconfig
-from dataclasses import dataclass
-from importlib.machinery import FileFinder
-from pathlib import Path
-from pkgutil import iter_modules
-from typing import Any
+from importlib import metadata
+from typing import TypedDict
 
 
-def _get_stdlib_modules() -> set[str]:
-    stdlib_path = sysconfig.get_paths()["stdlib"]
-    stdlib_modules = set(sys.builtin_module_names)  # start with built-ins
-
-    for root, _, files in os.walk(stdlib_path):
-        for filename in files:
-            if filename.endswith((".py", ".pyc", ".so", ".pyd")):
-                rel_path = os.path.relpath(os.path.join(root, filename), stdlib_path)
-                parts = rel_path.split(os.sep)
-
-                # Only consider top-level modules/packages
-                if parts:
-                    module = parts[0]
-                    if module.endswith((".py", ".so", ".pyd", ".pyc")):
-                        module = os.path.splitext(module)[0]
-                    stdlib_modules.add(module)
-
-    return stdlib_modules
+class IntrospectionResult(TypedDict):
+    top_level_stdlib_module_names: list[str]
+    distribution_names: list[str]
+    path: list[str]
 
 
-@dataclass(frozen=True, slots=True)
-class Module:
-    name: str
-    """The name of the module you'd `import` in a Python script"""
+def introspect_current_environment() -> IntrospectionResult:
+    # This is only available since Python 3.10!
+    top_level_stdlib_module_names = list(sys.stdlib_module_names)
 
-    distribution_names: set[str]
-    """
-    The name used to install this, e.g. 'pillow' instead of 'PIL'.
+    # This acts as a safety mechanism. We should definitely find information
+    # for all of them somewhere later.
+    distribution_names = list(
+        {dist.metadata["Name"] for dist in metadata.distributions()}
+    )
 
-    Can be multiple in case of a namespace package.
-    """
-
-    belongs_to_stdlib: bool
-
-    location: Path | None
-
-    def to_json_dict(self) -> dict[str, Any]:
-        return {
-            "name": self.name,
-            "distribution_names": list(self.distribution_names),
-            "belongs_to_stdlib": self.belongs_to_stdlib,
-            "location": str(self.location) if self.location is not None else None,
-        }
-
-    @staticmethod
-    def from_json_dict(value: dict[str, Any]) -> "Module":
-        return Module(
-            name=str(value["name"]),
-            distribution_names=set(value["distribution_names"]),
-            belongs_to_stdlib=bool(value["belongs_to_stdlib"]),
-            location=Path(value["location"]) if value["location"] is not None else None,
-        )
+    return IntrospectionResult(
+        top_level_stdlib_module_names=top_level_stdlib_module_names,
+        distribution_names=distribution_names,
+        path=sys.path,
+    )
 
 
-def get_available_modules_by_name() -> dict[str, Module]:
-    module_to_distribution_map = _build_module_to_distribution_map()
-    stdlib_modules = _get_stdlib_modules()
-
-    return {
-        module.name: Module(
-            name=module.name,
-            distribution_names=module_to_distribution_map.get(module.name, set()),
-            location=Path(module.module_finder.path)
-            if isinstance(module.module_finder, FileFinder)
-            else None,
-            belongs_to_stdlib=module.name in stdlib_modules,
-        )
-        for module in iter_modules()
-    }
+def print_serialized_introspection_result() -> None:
+    """This is what used when introspecting another env"""
+    print(json.dumps(introspect_current_environment()))